pipelines/OCR_deepseek/requirements.txt
google-cloud-aiplatform>=1.38.0
vertexai>=0.0.1
python-dotenv>=1.0.0

Pillow>=10.0.0  # For image processing
PyMuPDF>=1.23.0  # For PDF to image conversion (fitz)
pandas>=2.0.0  # For error logging CSV
